Revert firebase config
