+++ SmartClothingApp/src/components/visualizations/ActivityRings/ActivityRings.jsx
@@ -1,4 +1,8 @@
-import { Canvas, Fill, vec } from "@shopify/react-native-skia";
+import {
+  Canvas,
+  Fill,
+  vec,
+} from "@shopify/react-native-skia";
 import React from "react";
 
 import Ring from "./Ring";
